Convert layout.js to typescript
